refactor(shared): インフラ層を専用パッケージに分離
shared パッケージから Kafka、KurrentDB、TypeORM 関連のコードを
専用パッケージに分離しました。

## 新規パッケージ
- @system-board/shared-kafka: Kafka イベントパブリッシャー
- @system-board/shared-kurrentdb: KurrentDB イベントストアアダプター
- @system-board/shared-typeorm: TypeORM ユーティリティと ProcessedEvent エンティティ

## 変更内容
- Kafka 関連コードを shared-kafka パッケージに移動
- KurrentDB 関連コードを shared-kurrentdb パッケージに移動
- TypeORM 関連コードを shared-typeorm パッケージに移動
- shared パッケージの依存関係を整理
