+++ apps/packages/shared-kurrentdb/src/infrastructure/index.ts
@@ -1,0 +1,3 @@
+export * from './kurrentdb/KurrentConfig';
+export * from './kurrentdb/KurrentEventStoreAdapter';
+export * from './kurrentdb/KurrentModule';
