+++ apps/packages/shared-kurrentdb/src/infrastructure/kurrentdb/KurrentConfig.ts
