+++ apps/packages/shared-kurrentdb/src/infrastructure/kurrentdb/KurrentEventStoreAdapter.spec.ts
