+++ apps/packages/shared-kurrentdb/src/infrastructure/kurrentdb/KurrentModule.ts
