+++ apps/packages/shared-typeorm/src/infrastructure/index.ts
@@ -1,0 +1,4 @@
+export * from './typeorm/TypeOrmProcessedEventService';
+export * from './typeorm/entities/ProcessedEvent.entity';
+export * from './typeorm/utils/TypeOrmErrorHandler';
+export * from './typeorm/utils/TypeormErrorHandlerProvider';
