+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/entities/ProcessedEvent.entity.ts
