+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/utils/BaseTypeOrmErrorHandler.ts
