+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/utils/MySqlErrorHandler.ts
