+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/utils/PostgresErrorHandler.ts
