+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/utils/SqliteErrorHandler.ts
