+++ apps/packages/shared-typeorm/src/infrastructure/typeorm/utils/TypeOrmErrorHandler.ts
