+++ apps/packages/shared/src/infrastructure/index.ts
@@ -1,13 +1,5 @@
-// Kafka Infrastructure
-export * from './kafka/KafkaEventPublisher';
-export * from './kafka/KafkaModule';
-
 // Event Store Infrastructure
 export * from './eventstore/EventPersistenceService';
 export type { EventSerializer } from './eventstore/EventSerializer';
 export * from './eventstore/EventStore';
 export * from './eventstore/EventSerializerRegistry';
-export * from './eventstore/EventStoreModule';
-
-// PostgreSQL Infrastructure
-export * from './typeorm/TypeOrmProcessedEventService';
