+++ package.json
@@ -8,6 +8,9 @@
     "lint": "turbo lint",
     "format": "turbo format",
     "shared": "pnpm --filter '@system-board/shared'",
+    "shared-kafka": "pnpm --filter '@system-board/shared-kafka'",
+    "shared-kurrentdb": "pnpm --filter '@system-board/shared-kurrentdb'",
+    "shared-typeorm": "pnpm --filter '@system-board/shared-typeorm'",
     "system-management": "pnpm --filter '@system-board/system-management'",
     "vulnerability-management": "pnpm --filter '@system-board/vulnerability-management'",
     "relationship-management": "pnpm --filter '@system-board/relationship-management'",
